Implement ClientActionHandler

+++ src/main/kotlin/adapter/in/http/client/ClientInputDispatcher.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.channels.SendChannel
 class ClientInputDispatcher(private val channel: SendChannel<ClientAction>) {
     suspend fun dispatchToChannel(clientId: String, message: String) {
         if (NumberValidator.validateNumberIsCorrect(message)) {
-            return channel.send(ClientAction.NewValue(message))
+            return channel.send(ClientAction.NewValue(message.toInt()))
         }
         if (message == "terminate") {
             return channel.send(ClientAction.Shutdown)
